modeling/requirements.txt
absl-py==2.0.0
aiofiles==22.1.0
aiosqlite==0.19.0
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
databricks-cli==0.18.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker==6.1.3
entrypoints==0.4
executing==2.0.1
fastapi==0.109.0
fastjsonschema==2.19.1
filelock==3.13.1
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.47.0
fqdn==1.5.1
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==20.1.0
h5py==3.10.0
hub-sdk==0.0.3
idna==3.6
imageio==2.33.1
importlib-metadata==6.11.0
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.20.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.7.1
jupyter_server==2.12.2
jupyter_server_fileid==0.9.1
jupyter_server_terminals==0.5.1
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mlflow==2.4.1
mpmath==1.3.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.0.1
pandocfilters==1.5.0
parsenvy==3.0.2
parso==0.8.3
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
querystring-parser==1.2.4
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.11.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.35.1
statsmodels==0.14.0
streamlit==1.30.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io==0.34.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
terminado==0.18.0
testbook==0.4.2
tf-slim==1.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ultralytics==8.1.1
uri-template==1.3.0
urllib3==2.1.0
utils==1.0.2
validators==0.22.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.14.1
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.17.0
